Move home components into subfolder
