+++ app/features/home/__tests__/HomePage.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, within, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import HomePage from '@/app/features/home/HomePage';
+import HomePage from '@/app/features/home/components/HomePage';
 import { ThemeProvider } from '@/app/providers/ThemeContext';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
 import { Verse } from '@/types';
@@ -25,7 +25,7 @@ jest.mock('next/link', () => {
 });
 
 // Mock VerseOfDay to avoid fetch during tests
-jest.mock('@/app/features/home/VerseOfDay', () => () => <div>VerseOfDay</div>);
+jest.mock('@/app/features/home/components/VerseOfDay', () => () => <div>VerseOfDay</div>);
 
 beforeAll(() => {
   Object.defineProperty(window, 'matchMedia', {
