+++ app/features/home/components/HomePageBackground.tsx
