+++ app/features/home/components/VerseOfDay.tsx
