+++ app/features/home/page.tsx
@@ -1,4 +1,4 @@
-import HomePage from './HomePage';
+import HomePage from './components/HomePage';
 
 export default function Page() {
   return <HomePage />;
